Hashtable: Change size of Alc_Hash from 32 to 64 bits.

+++ include/alc/hashtable.h
@@ -4,7 +4,7 @@
 #include <alc/defs.h>
 
 typedef u8 Alc_Control;
-typedef u32 Alc_Hash;
+typedef u64 Alc_Hash;
 
 typedef struct {
   Alc_Control *control_block;
